added translation page
