+++ final_project/src/main/webapp/index.html
@@ -17,6 +17,17 @@
     <h1>FAMA Charity</h1>
    <span style="font-size:30px;cursor:pointer" onclick="openNav()">&#9776; open</span>
     <hr>
+       <form action="/translatePage" method="Post">
+    <select name="languageCode" >
+       <option value="en">English</option>
+      <option  value="zh">Chinese</option>
+      <option  value="es">Spanish</option>
+      <option  value="hi">Hindi</option>
+      <option  value="ar">Arabic</option>
+     </select>
+    
+    <button onclick="getCauses()"> Translate</button>
+        </form>
     <!-- sidenav -->
     <div id="mySidenav" class="sidenav">
       <a href="javascript:void(0)" class="closebtn" onclick="closeNav()">&times;</a>
@@ -80,12 +91,82 @@
         Our Causes:
         <hr>
       </h3>
+
         <div id= "posts-container">
             <div class="row">
-               <div id = "post-container-column" class="column">
-               </div>
+             <div id = "post-container-column" class="column">
+             </div>
           </div>
         </div>
+
+      <!-- <div class="row">
+        <div class="column">
+          <div class="content">
+          </div>
+        </div> -->
+
+
+      <!-- <div class="row">
+        <div class="column">
+          <div class="content">
+            <a href="/yemen.html">
+            <img id="img" src="./images/yemen.jpg" style="width: 100%; height: 100%;"></a>
+            <h3 id="title"> </h3>
+            <p id="discription" ></p>
+          </div>
+        </div> -->
+        <!-- <div class="column">
+          <div class="content">
+            <a href="/yemen.html">
+            <img id="img" src="./images/yemen.jpg" style="width: 100%; height: 100%;"></a>
+            <h3> The World’s Worst Humanitarian Crisis</h3>
+            <p>Called the “world’s worst humanitarian crisis,” in Yemen, more than 20 million people are at risk for famine, and aid agencies are desperately funneling as much aid as possible to at-risk families.</p>
+          </div> -->
+       <!-- </div>
+
+
+
+        <div class="column">
+          <div class="content">
+            <a href="/yemen.html">
+            <img id="img" src="./images/yemen.jpg" id="img" style="width: 100%; height: 100%;"></a>
+            <h3 id="title"> </h3>
+            <p id="discription">Called the “world’s worst humanitarian crisis,” in Yemen, more than 20 million people are at risk for famine, and aid agencies are desperately funneling as much aid as possible to at-risk families.</p>
+          </div>
+        </div>
+
+
+
+       <div class="column">
+          <div class="content">
+            <a href="/yemen.html">
+            <img id="img" src="./images/yemen.jpg" style="width: 100%; height: 100%;"></a>
+            <h3> The World’s Worst Humanitarian Crisis</h3>
+            <p>Called the “world’s worst humanitarian crisis,” in Yemen, more than 20 million people are at risk for famine, and aid agencies are desperately funneling as much aid as possible to at-risk families.</p>
+          </div>
+        </div>
+       <div class="column">
+          <div class="content">
+            <a href="/yemen.html">
+            <img id="img" src="./images/yemen.jpg" style="width: 100%; height: 100%;"></a>
+            <h3> The World’s Worst Humanitarian Crisis</h3>
+            <p>Called the “world’s worst humanitarian crisis,” in Yemen, more than 20 million people are at risk for famine, and aid agencies are desperately funneling as much aid as possible to at-risk families.</p>
+          </div>
+        </div>
+     <div class="column">
+          <div class="content">
+            <a href="/yemen.html">
+            <img id="img" src="./images/yemen.jpg" style="width: 100%; height: 100%;"></a>
+            <h3> The World’s Worst Humanitarian Crisis</h3>
+            <p>Called the “world’s worst humanitarian crisis,” in Yemen, more than 20 million people are at risk for famine, and aid agencies are desperately funneling as much aid as possible to at-risk families.</p>
+          </div>
+        </div>
+      </div>
+      -->
+    </div>
+
+
+
     <div id="donate" >
         <h3>
         Donate:
@@ -188,8 +269,6 @@
           </p>
           <p>FAMA Charity &copy; 2020</p>
         </div>
-
-        
       </footer>
     </div>
   </body>
